A proof-of-concept Node implementation
I haven't tested it extensively, but types seem to mostly check out,
and some simple (branching!) pipelines even work as expected.  Type
inference is, as expected, a wreck because of all the type families,
but it seems to generally work fine if the shapes (i.e. number of
connections) of the nodes are known, which seems like good enough for
practical usage.

I have no idea how performance compares with that of other streaming
libraries; I'd expect it to be slower, but not by much if everything
is properly specialised.