Add ability to fuse open connections
This was actually surprisingly easy.

I think this will be the extent of the support for fusing connections
in this library.

Another attractive alternative is automatic fusing based on the
type (so, switch from HSum to something like Union).  A use case is
easy to imagine: for example, a usual linear pipeline, but each step
also produces some values for the log or something.  With `fuseCons`,
user has to fuse all these extra connections manually, which can be
tedious and boring; automated fusing would relieve them of that duty.

However, doing things automatically tends to backfire in surprising
ways at the most inopportune moments (“explicit is better than
implicit”), not to mention the problems that arise when one *wants* to
have multiple identical ports (see `tee`).  That would require some
pointless newtype wrappers to distinguish the two, which is bad enough
by itself, but what happens on connection?  Either we drop fundeps
from Annihilate (and say goodbye to what little type inference we've
got left), or we now have to wrap *other* end in newtype wrappers as
well!  And what if it already has some wrappers of its own?  Oh, the
horror.  No, better do things manually.