Add pipe-like streaming
With producers connecting to pipes connecting to consumers in a linear
fashion.

The reason for this type machinery is that, while the Yielding
connection is almost always second (and so connectOn t1 t0 almost
always works), *producers* have Yielding as their *one and only*
connection, so t0.  As a neat side effect, we can now sandwich
additional connections in the middle, and deal with them later.