Add All type family to provide cleaner context for Functor (HSum)
Cleaner instance, cleaner contexts on functions, what not to like?

It's not a secret that this change is inspired by type-combinators
package; the next order of business is to figure out how much effort
would it take to transition to that from home-brewn manipulations, and
how much will I still have to implement on top of it.