Replicate HSum functionality based on type-combinators
Type families have mostly turned into classes with functional
dependencies, and we'll have to see how that works out, but other than
that, I'm *incredibly* happy with this.

Even though I was unable to figure out how to replicate most of
functionality used based on type-combinators built-in methods, even
just writing code in what I perceive to be type-combinators style has
lead to vastly improved contexts and more sensible implementations.
Instead of weird implementation-leaking classes of the HSum module, we
now have only two custom classes, both of them having to do with the
general structure of the lists, as opposed to being specific for the
problem at hand.

I'm also happy with the new method of denoting indices, as supplying
actual index as an argument now destructures the lists enough that the
type checker can actually resolve other contexts!  Hooray!