Add a PrimMonad instance
I haven't really used it myself, but it seems to be used for mutable
containers of all sorts, which sure sounds useful.

I could not quite discern the status of the primitive package (it does
not belong to core libraries, but is managed by Core Libraries
Committee or something along these lines), but it seems official
enough to be a benign addition to the dependency list.