Add a forsOnLeaf family of functions
Do they supersede various mapsOnM-family functions?  That is an
interesting question, which I view as a benchmark of the library's
optimisation capabilities; which is to say, if they do not because of
performance hits, then I need more INLINE and RULES pragmas (of which
I don't currently have any, but they're coming).