Replace cata from recursion-schemes with a hand-rolled version
This has several advantages:

1) Better control wrt. inlining, possible rules, etc.

2) Less dependencies!  I didn't need /that/ much from
recursion-schemes anyway.

3) Better encapsulation.  I was a little bit worried about the
Recursive instance of Node, as it exposed a little bit too much of
implementation outside the Internal module.  Sure, the constructors of
the Base functor were hidden, but there was this lingering worry that
someone might somehow exploit the provided argument (it's still
possible to distinguish bottom, for example!) to observe violation of
monad transformer laws.  Well, not anymore.