Add yield
I am /so/ not sure about this function.  On the one hand, I want it to
be usable both from producers and from pipes, which necessitates this
complicated type.  On the other, this is weird, makes it usable in
more circumstances than I'd want to.  Also, should 'each' therefore
also have similar type?