Explain distribute to myself
It turns out that 'distribute' is essentially 'run' in disguise (in
fact, it is implemented that way in pipes).  I have considered making
the following function, which factors out the common structure:

unsafeInterpret ::
     (Functor m, Monad n, All Functor cs)
  => (forall x. m x -> n x)
  -> (forall x. FSum cs x -> n x)
  -> Node cs m r
  -> n r
unsafeInterpret e c = cata alg
  where
    alg (ReturnF r) = pure r
    alg (EffectF eff) = join (e eff)
    alg (ConnectF con) = join (c con)

and then

run = unsafeInterpret id (absurd . nilFSum)
distribute = unsafeInterpret (hoist lift) (lift . connects)

but if anything, this seems /more/ mysterious, as the names of the
constructors are gone, and matching passed functions to them requires
noticeable mental effort.  Better large function that is easier to
follow, than one-liner that is difficult to understand.

As I was writing it, I realised that this is just 'foldFree' from
‘free’ package.