Flesh out Fold
I am not sure whether everything is strict enough—I suppose benchmarks
should tell—but seems fine so far.

I had a long list of specific folds (sum, product, all, that stuff)
but eventually deleted them; they're kind of trivial to write by hand
if needed, are already provided by foldl package if which this module
strongly hints about, and with four variants of each there were a
/lot/ of them, which looked daunting in the docs, caused a lot of
repetition in the code, and caused me to procrastinate for
impressively long.  Screw all that.