Flesh out Link module
I have switched to more pipe-esque operators.  They seem to look
slightly better than the |$ mess.  But whatever; the most benefit is
that it provides some pre-established conventions for duplex
communcation.

Speaking of which, there's duplex communication now!  I mostly added
it for feature parity with pipes, and because it helps establish
unambiguous control flow.  However, there's no documentation on these
currently, because supporting module does not exist yet.

All actual heavy-lifting happens in genericLinkOn; it remains to be
seen whether liberal inlining will be able to bring performance of
this variant on par with hand-written operators, but I sure hope so!