Add forsAll function
Note the new implementation of 'distribute', which now doesn't use
internal constructors and, sans unsafeHoist, is implemented entirely
via public interface.  A slight worry is that the new version
traverses the Node /three/ times, as opposed to one in the old one,
but that shouldn't be much of a problem in practice.  In most cases,
distribute probably won't be called very often, and for resource
management (coming!) the computation is probably IO-bound anyway.

The new implementation now closely resembles the one in pipes.