Flesh out Trans module
Writer is conspicuously missing.  There are two reasons for that:

1. Node currently does not have MonadWriter instance anyway.  I have
softened to the idea of having it, but not right now.

2. Monadic bind in WriterT builds up a left-associated chain of binds
in the base monad, which means quadratic slowdown for free base
monads (such as Node).  This means that I either roll out a custom
runner for WriterT, or switch to codensity representation for Node.  I
was planning to investigate the second option anyway, so I'll wait
until I have a definite opinion on that.