Add Resource module
The problem of safe resource management is 80% solved by ResourceT, so
thanks to Michael Snoyman for that.  The only remaining problem is
prompt release of resources acquired by interrupted nodes.  Conduit
solves that by using finalizers, but they are a pain to work with (and
are said to complicate 'for' semantics), so I've come up with this
clever(?) region-based trick instead, which I rather like.

Unfortunately, the desire to handle as many exit paths as possible has
led to combinatorial explosion of functions (luckily, limited to 4 per
family); I'd like to know a better way if there is one, but I don't,
so here we are.