Drop NodeF and make Node explicitly recursive
The main motivation here is that the presence of extra newtype layers
seems to make rewrite rules a bit more fragile; in particular, when I
tried to make old bind_ operate on Nodes and match on (Node (SmthF
...)) in rules, they didn't fire.

As for the overall result, I'm not sure.  Using cata/transform now
leads to slightly more mysterious code (because the constructor names
are gone), but they (thankfully) aren't used that often anyway.
Direct pattern-matches of Node are now a bit cleaner.