Add an Inductive class
The idea here is that now the “recursive” calls aren't technically
recursive anymore, as they belong to a different instance.  The net
effect is that GHC should be able to unroll the loops for known
loops.  Performance in non-unrolled situations will probably suffer
due to all the virtual calls through dictionaries…

Plus, every function will get weird Inductive constraint.  Also, the
class itself is huge and ugly.  On balance, though, it is still better
than what we have now, so I think this is a good tradeoff.

I suspect the idea itself is not new, but I have borrowed it from Oleg
Genrus's fin package.

This particular commit only introduces a class, the major rewrite will
come later.

Note that routing all induction through one class led to an ugly wart
by the name injectReplaced. I'm sorry about that, but I still believe
that to be improvement over status quo.