Change all combinators to use Inductive
Making Inductive a superclass of all other classes defined in
Combinators is kind of bad, since it reduces their generality, but
helps reduce the required number of constraints in other places, thus
simplifying public interface.  I think this is a good tradeoff.

Somewhere in the process, rewrite rules for appendAll ceased to
typecheck.  I'm not too worried about them, since the plan is to get
rid of it anyway.  This is also the general reason not to have lemmata
in Inductive itself — I expect some major changes on that front.