Switch Node to a RWR-inspired representation
Just in case, RWR here refers to “reflection without
remorse” (http://okmij.org/ftp/Haskell/zseq.pdf).

The basic idea here is that having quadratic complexity for operations
that clearly don't require it is unacceptable.  This applies both to
left-associated binds and reflection (as represented by 'inspect'
family of functions).  Hence RWR, which offers amortized O(1) bind and
reflection.

Unsurprisingly, such significant changes to the core data type of the
library impact pretty much all of its code, so this is going to be a
long commit message.

First of all, the choice of internal representation.  Okasaki's book
has a more sophisticated catenable queue, which, presumably, maintains
O(1) even in presence of sharing, but since Nodes should really be
consumed in an affine fashion, binary tree is good enough.  Both
monad-skeleton and freer-effects use it, so I hope they have verified
its complexity properties, because I haven't.

I have currently used Sum to represent both effects and connections in
one Node constructor.  I *may* split it into two constructors in the
future if that provides any noticeable performance improvement.
Pretty much no one interacts with Node's constructors directly after
these changes, so it should be easy to do.

Since I was rewriting a significant chunk of the library, I decided to
hop on this opportunity and lay foundations for future fusion work in
terms of 'build'.  An unfortunate side effect is that everything is
now written in a horrible unreadable CPS style, with manual
chaining (I send special rays of hatred to scanM and zipWith).  I'm
sorry, it's for the greater good.

I don't like new liftCatch one bit, but I don't have any better ideas
either.

I have dropped transformCons, because it was rather useless and didn't
really fit with the whole fusion thing.  Speaking of which, I also
dropped various map..M functions, because they're a pain to write and
for* family of functions is strictly more general anyway.

For inspect* family of functions I decided to use Coyoneda.  Strictly
speaking, everything is still made of functors, so I could've retained
the old interface, but that is changing soon, and I wanted to switch
to the whole Freer idea in one swoop.  As for specifically importing
Coyoneda from kan-extensions, it felt silly to introduce own type for
such a well-known and well-understood thing.  On the other hand,
depending on a whole extra library just for the sake of one simple
type also feels kind of less silly, though probably less so.

gforAll currently does more work than is strictly necessary (there's
no actual need to traverse the loop body anymore), but that will
hopefully change with future fusion work.

The most obvious user-facing change is probably significant reduction
in constraints.  I am *so* happy about that, especially since haddock
is notoriously bad at handling long constraint lists.

Speaking of which, pretty much all of documentation is now out-of-date
and wrong.  Since this isn't the last major change planned for the
near future, I'd rather postpone that problem for now.

And yes, the performance did go down the drain.  One of the prominent
directions for future work at this point is recovering as much of as
is possible.