Introduce the foundations of the fusion framework
There's a lot of ridiculous things going on with this commit, so let
me explain.

First of all, my initial goal was to get rid of `Cat` constructors on
simple right-associated binds, similarly to how the rules on `bind_`
worked before the switch to RWR.  One thing led to another, and it
turned out that the only way I can make this happen is to introduce a
full-blown cata/build fusion, which, I admit, was always planned, but
not quite so soon.

The huge downside of the resulting solution is that bind_ is now
promptly rewritten into a loopy version, reintroducing *exactly* the
quadratic complexity that I was trying to get rid of.  The reverse
rewrite rule ‘uncata’, is supposed to help with that, but I haven't
tested it yet.

Speaking of which, the whole rewrite rule machinery is surprisingly
finicky; as mentioned in comments, the only way to make it work seems
to involve reducing the rank of the types involved via Coyoneda.
Furthermore, the existence of the fusion rules seems to interfere with
inlining and specialisation in some bizarre ways, so that without the
bindConst_ the performance is easily three times worse.