Add a simple benchmark for left-associated binds
This is not a super-scientific comparative benchmark like I've tried
to do elsewhere, just a quick check that the 'uncata' rule works as
intended, and the complexity on left binds is indeed linear.  I'm
happy to report that that is indeed the case; however, the overall
performance is still ~27–28 times worse than on right-associated
binds, so…  Well, at least the asymptotic behaviour is correct.