Split constructors for connections and effects
This seems to give a ~10% performance boost; not too bad.  On the
'sum' benchmark cobweb is now roughly on par with pipes.

A separate interesting point is that now there's some freedom to make
use of monadic nature of 'm' in the effect constructor.  I'm not too
keen on doing things, since this will complicate things and
reintroduce constraints on 'm' on most of the functions, which is
probably morally correct, but somehow unpleasant.