Add an inspection-testing based test-suite for fusion framework
Unfortunately, inspection-testing only got run time testing support in
0.2, and the latest version currently available via nix is 0.1.2.  So,
compile-time tests will have to do.

And already the tests have uncovered a problem.  AFAICT, the reason
here is that, in absence of any pragmas, GHC inlines an optimised
version of the function — which in this case involves inlining build,
and thus breaking fusion.  Looking around, there are a lot of
functions that could benefit from that not happening, which means, in
practice, that there will soon be a lot more INLINE pragmas throughout
the library.