Remove unsafeHoist, changing regular hoist to its behaviour
The new version is potentially unsafe if the function passed is not a
monad morphism, but…

1. The docs for mmorph kind of hint at, but never quite spell out as
explicitly as I'd like, that that should never be the case.

2. The new version is faster in general, but especially so given the
whole fusion thing.  I'm not sure it is worth abandoning that
performance for the sake of a rare and questionably legal use case.

3. The docs for mmorph never actually state that as a
law (surprisingly), but, if we take the relationship between normal
Functor and Monad as an example, we arrive at the following reasonable
law:

  embed (lift . f) = hoist f

Well, 'embed' pretty much has no choice but to expose the violation of
the monad transformer laws, given that the monad constraint is on the
other side, so it's impossible to squash neighbouring effects.  So… if
we can tolerate that, I say we tolerate hoist doing the same,
especially if it comes with other benefits.