Extend fusion testing to the entire Internal module
And, in the process, make the fusion actually work.

Besides the numerous INLINE pragmas, the least obvious change is
CONLIKE pragmas on build_ and augment.  Without them, the test for
local (and possibly others), fails.  I'm not sure why exactly (the
test does not appear to explicitly share monadic actions), but
whatever.

One possible problem is these are a lie of sorts; CONLIKE is supposed
to mean 'cheap, like a constructor', and build_ can be much more
expensive like that, so that duplicating it can significantly increase
the computations done.

Luckily, though, sharing values of streaming monads is generally an
undesired behaviour anyway; the whole point of the streaming library
is that nothing remains in memory for longer than it absolutely has
to.  In this light, the fusion problems are a blessing in disguise, as
they nudge gently in the direction of forcing the compiler to reduce
sharing.